Move extensions to root SerilogTracing namespace (as done in Serilog); split OTLP logs and traces sinks
