+++ src/SerilogTracing.Instrumentation.AspNetCore/SerilogTracing.Instrumentation.AspNetCore.csproj
@@ -4,6 +4,7 @@
         <TargetFramework>net8.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
+        <RootNamespace>SerilogTracing</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
