+++ src/SerilogTracing.Sinks.OpenTelemetry/Collections/EnumerableExtensions.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Serilog.Collections;
+namespace SerilogTracing.Collections;
 
 static class EnumerableExtensions
 {
