+++ src/SerilogTracing.Sinks.OpenTelemetry/SerilogTracing.Sinks.OpenTelemetry.csproj
@@ -13,7 +13,7 @@
 		<RepositoryUrl>https://github.com/serilog/serilog-sinks-opentelemetry</RepositoryUrl>
 		<RepositoryType>git</RepositoryType>
 		<GenerateDocumentationFile>true</GenerateDocumentationFile>
-		<RootNamespace>Serilog</RootNamespace>
+		<RootNamespace>SerilogTracing</RootNamespace>
 		<PackageReadmeFile>README.md</PackageReadmeFile>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<Nullable>enable</Nullable>
