+++ src/SerilogTracing.Sinks.Seq/SerilogTracing.Sinks.Seq.csproj
@@ -2,6 +2,7 @@
 
     <PropertyGroup>
         <TargetFramework>net6.0</TargetFramework>
+        <RootNamespace>SerilogTracing</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
