+++ src/SerilogTracing.Sinks.Zipkin/SerilogTracing.Sinks.Zipkin.csproj
@@ -2,6 +2,7 @@
 
     <PropertyGroup>
         <TargetFramework>net6.0</TargetFramework>
+        <RootNamespace>SerilogTracing</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
