+++ src/SerilogTracing.Sinks.Zipkin/Sinks/Zipkin/ZipkinSink.cs
